fix circular import by ditching separate file for mirostat
